Moved admin-links from top-level to catalog sub-menu
There is no reason to have this as it's own top-level menu since it
clearly has to do solely with products and that is directly under
catalog. This also helps to keep the admin clean and uncluttered.

Also removed the link to attachment config, since no other part of
Magento's main menu include links to the config area for each fature.